src/ghostpost/serializers.py
from rest_framework import serializers
from .models import Post

class PostSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Post
        fields = ['id', 'u_id', 'post_type', 'text', 'up', 'down', 'date']