pyproject.toml
[tool.poetry]
name = "gp_backend"
version = "0.1.0"
description = ""
authors = ["Tamoya Hopkins <tamoyashopkins@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.2"
ipdb = "^0.12.3"
djangorestframework = "^3.11.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
